Builder pattern added
