+++ src/main/java/flight/reservation/Customer.java
@@ -3,10 +3,10 @@ package flight.reservation;
 import flight.reservation.flight.ScheduledFlight;
 import flight.reservation.order.FlightOrder;
 import flight.reservation.order.Order;
+import flight.reservation.order.FlightOrderBuilder;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Customer {
 
@@ -21,36 +21,32 @@ public class Customer {
     }
 
     public FlightOrder createOrder(List<String> passengerNames, List<ScheduledFlight> flights, double price) {
-        if (!isOrderValid(passengerNames, flights)) {
-            throw new IllegalStateException("Order is not valid");
-        }
-        FlightOrder order = new FlightOrder(flights);
-        order.setCustomer(this);
-        order.setPrice(price);
-        List<Passenger> passengers = passengerNames
-                .stream()
-                .map(Passenger::new)
-                .collect(Collectors.toList());
-        order.setPassengers(passengers);
-        order.getScheduledFlights().forEach(scheduledFlight -> scheduledFlight.addPassengers(passengers));
+        FlightOrder order = new FlightOrderBuilder(flights)
+                .setCustomer(this)
+                .setPrice(price)
+                .setPassengers(passengerNames, flights) // Validation happens here
+                .build();
+
         orders.add(order);
         return order;
     }
 
-    private boolean isOrderValid(List<String> passengerNames, List<ScheduledFlight> flights) {
-        boolean valid = true;
-        valid = valid && !FlightOrder.getNoFlyList().contains(this.getName());
-        valid = valid && passengerNames.stream().noneMatch(passenger -> FlightOrder.getNoFlyList().contains(passenger));
-        valid = valid && flights.stream().allMatch(scheduledFlight -> {
-            try {
-                return scheduledFlight.getAvailableCapacity() >= passengerNames.size();
-            } catch (NoSuchFieldException e) {
-                e.printStackTrace();
-                return false;
-            }
-        });
-        return valid;
-    }
+    // private boolean isOrderValid(List<String> passengerNames,
+    // List<ScheduledFlight> flights) {
+    // boolean valid = true;
+    // valid = valid && !FlightOrder.getNoFlyList().contains(this.getName());
+    // valid = valid && passengerNames.stream().noneMatch(passenger ->
+    // FlightOrder.getNoFlyList().contains(passenger));
+    // valid = valid && flights.stream().allMatch(scheduledFlight -> {
+    // try {
+    // return scheduledFlight.getAvailableCapacity() >= passengerNames.size();
+    // } catch (NoSuchFieldException e) {
+    // e.printStackTrace();
+    // return false;
+    // }
+    // });
+    // return valid;
+    // }
 
     public String getEmail() {
         return email;
@@ -70,6 +66,7 @@ public class Customer {
 
     public List<Order> getOrders() {
         return orders;
+
     }
 
     public void setOrders(List<Order> orders) {
